Declare DuckyEncode::pending_ctx_size_ as size_t
Declare DuckyEncode::pending_ctx_size_ as size_t instead of int. Fix the
following MSVC warning:
av1\ducky_encode.cc(551,43): warning C4267: '+=': conversion from
'size_t' to 'int', possible loss of data

Include <cstddef> for size_t and <cstdint> for uint8_t and int64_t. See
the Style Guide:
https://google.github.io/styleguide/cppguide.html#Include_What_You_Use

List C++ system headers in alphabetical order. See the Style Guide:
  Within each section the includes should be ordered alphabetically.
https://google.github.io/styleguide/cppguide.html#Names_and_Order_of_Includes

+++ av1/ducky_encode.h
@@ -12,9 +12,11 @@
 #ifndef AOM_AV1_DUCKY_ENCODE_H_
 #define AOM_AV1_DUCKY_ENCODE_H_
 
+#include <cstddef>
+#include <cstdint>
+#include <memory>
 #include <string>
 #include <vector>
-#include <memory>
 
 #include "aom/aom_encoder.h"
 #include "av1/encoder/firstpass.h"
@@ -90,7 +92,7 @@ class DuckyEncode {
   std::unique_ptr<EncodeImpl> impl_ptr_;
   bool write_temp_delimiter_;
   std::vector<uint8_t> bitstream_buf_;
-  int pending_ctx_size_;
+  size_t pending_ctx_size_;
 };
 }  // namespace aom
